// src/api/generated/headscale.ts
export interface components {
    schemas: {
// ...
        v1User: {
            /** Format: uint64 */
            id?: string;
            name?: string;
            /** Format: date-time */
            createdAt?: string;
            displayName?: string;
            email?: string;
            providerId?: string;
            provider?: string;
            profilePicUrl?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}